Update 20240203 - Database Implementation
Implemented the database table for the following:

- Roles
- Users
- Password Reset
- Failed Jobs
- Personal Access Token
- Categories
- Products
- Inventories
- Product Categories
- Deliveries
- Transactions
- Stock Orders
- Supplies

Furthermore, these tables are split into two database defined under the config/database.php.

The project is now only composed of the contents of SuppandDeliCOMETAL directory, removing the Supply-and-Delivery directory and its unused contents.

Lastly, a GitHub workflow is added to run tests on the Laravel project.