Update 20240203 - Backend Rehaul
Rehauled a lot of backend stuff and refactored a lot of routing.

The routing is now categorized and grouped with their respective subject while moving their controller functions to their respective controller files.

Laravel Sanctum is now also enabled to all users logging in to enable their accounts to use the REST APIs.

Various migrations are also updated to make sense by moving their columns to the other table.

Validation are added to cleanse the input users are providing, adding a bit more security to the web application.

A JS code block was also commented out to allow the forms to work again.

Login page is now the designated landing page for the web application.